--- deprecated/grepCloneTest.txt
@@ -1,4 +1,0 @@
-From 707b1b905ffdfa16b508c420c28596bd424f90a6 Mon Sep 17 00:00:00 2001
-From: Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
-Date: Wed, 10 Feb 2021 06:38:15 -0800
-Subject: [PATCH 1/3] Manual kustomize update.
